+++ Common/src/main/java/de/markusbordihn/adaptiveperformancetweaks/feature/benchmark/scenario/XpScenario.java
@@ -52,6 +52,11 @@ public final class XpScenario implements BenchmarkScenario {
   }
 
   @Override
+  public boolean shouldFacePlayerToFocus() {
+    return true;
+  }
+
+  @Override
   public void setup(BenchmarkScenarioContext context) {
     Vec3 center = context.center();